refactor(backends): split gitlab_sync.py into PR + issue modules (#522)
Extract 20 @classmethod methods (that never used cls) into module-level
functions across two new files:
- gitlab_sync_prs.py (270 LOC): PR building, upserting, discussion parsing
- gitlab_sync_issues.py (179 LOC): issue fetching, label resolution

gitlab_sync.py shrinks from 561 → 118 LOC as a thin orchestrator.
No behavioral changes — pure extraction.

Closes #522
